pyproject.toml
[tool.poetry]
name = "fhir-tx-client"
version = "0.1.0"
description = ""
authors = ["Axel Vanraes <axel.vanraes@tiro.health>"]
readme = "README.md"
packages = [{include = "fhir-tx-client"}]

[tool.poetry.dependencies]
python = "^3.10"
fhirpy = "^1.3.1"
fhir-resources = "^6.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
asyncio = "^3.4.3"
pytest-asyncio = "^0.20.3"
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
